make run more flexible and usable without CLI
