+++ ossdbs/__init__.py
@@ -4,8 +4,6 @@
 """Open-source software for deep brain stimulation."""
 import logging
 
-import ngsolve
-
 from ossdbs.api import (
     create_bounding_box,
     generate_brain_model,
@@ -22,29 +20,12 @@ from ossdbs.api import (
     set_contact_and_encapsulation_layer_properties,
 )
 from ossdbs.fem import ConductivityCF, Mesh
+from ossdbs.main import main_run, set_logger
 from ossdbs.model_geometry import BrainGeometry, ModelGeometry
 from ossdbs.utils.nifti1image import DiffusionTensorImage, MagneticResonanceImage
 
 _logger = logging.getLogger(__name__)
 _logger.addHandler(logging.NullHandler())
-
-
-def set_logger(level=logging.INFO):
-    """Set log level."""
-    _logger.setLevel(level)
-    if level == logging.DEBUG:
-        ngsolve.ngsglobals.msg_level = 10
-    # to avoid multiple output in Jupyter notebooks
-    if len(_logger.handlers) == 1:
-        ch = logging.StreamHandler()
-        ch.setFormatter(logging.Formatter(logging.BASIC_FORMAT))
-        ch.setLevel(level)
-        _logger.addHandler(ch)
-    else:
-        for handler in _logger.handlers:
-            if isinstance(handler, logging.StreamHandler):
-                handler.setFormatter(logging.Formatter(logging.BASIC_FORMAT))
-                handler.setLevel(level)
 
 
 __all__ = (
@@ -68,4 +49,5 @@ __all__ = (
     "Mesh",
     "run_volume_conductor_model",
     "generate_point_models",
+    "main_run",
 )
